move walkers into a sep dir
