+++ src/walker/index.ts
@@ -1,0 +1,2 @@
+export * from './walkOpenAPI';
+export * from './walkSwagger';
