+++ src/walker/types.ts
@@ -1,0 +1,9 @@
+import {LintError} from '../types';
+
+export type WalkerResult<T> =
+    | {
+          visitors: T;
+      }
+    | {
+          errors: LintError[];
+      };
